+++ src/import-exports/index.ts
@@ -2,5 +2,6 @@ export {IExportDefault, IExportDefaultOptions, create_export_default} from './ex
 export {IExportEqual, IExportEqualOptions, create_export_equal} from './export-equal';
 export {IExportNamed, IExportNamedOptions, create_export_named} from './export-named';
 export {IExportNamespace, IExportNamespaceOptions, create_export_namespace} from './export-namespace';
+export {IImportEqual, IImportEqualOptions, create_import_equal} from './import-equal';
 export {IImportNamed, IImportNamedOptions, create_import_named} from './import-named';
 export {IImportNamespace, IImportNamespaceOptions, create_import_namespace} from './import-namespace';